+++ package.json
@@ -1,10 +1,13 @@
 {
   "dependencies": {
     "@tanstack/react-query": "^5.64.1",
+    "autoprefixer": "^10.4.20",
     "axios": "^1.7.9",
     "js-cookie": "^3.0.5",
     "next-auth": "^4.24.11",
-    "react": "^19.0.0"
+    "postcss": "^8.5.1",
+    "react": "^19.0.0",
+    "tailwindcss": "^3.4.17"
   },
   "devDependencies": {
     "@types/js-cookie": "^3.0.6",
@@ -13,6 +16,13 @@
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-prettier": "^5.2.2",
-    "prettier": "^3.4.2"
+    "prettier": "^3.4.2",
+    "typescript": "5.7.3"
+  },
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   }
 }
