+++ src/app/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <div>
+      <h1>페이지</h1>
+    </div>
+  );
+}
